Split build battle into server-side decorator and client-side reward handler.
